# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients": {
      "post": {
        "operationId": "PatientsController_create",
        "summary": "Criar novo paciente",
        "description": "Cria um novo paciente com todas as informações necessárias.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Paciente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "409": {
            "description": "Paciente já existe com o mesmo CPF ou email"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PatientsController_findAll",
        "summary": "Listar todos os pacientes",
        "description": "Retorna uma lista de todos os pacientes cadastrados.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de pacientes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/patients/{id}": {
      "get": {
        "operationId": "PatientsController_findOne",
        "summary": "Buscar paciente por ID",
        "description": "Retorna os detalhes de um paciente específico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "summary": "Atualizar paciente",
        "description": "Atualiza as informações de um paciente existente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paciente atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Paciente não encontrado"
          },
          "409": {
            "description": "Conflito com outro paciente (CPF ou email já existente)"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PatientsController_remove",
        "summary": "Remover paciente",
        "description": "Remove um paciente do sistema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Paciente removido com sucesso"
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/patients/{id}/measurements": {
      "post": {
        "operationId": "PatientsController_createMeasurement",
        "summary": "Adicionar medição",
        "description": "Adiciona uma nova medição para um paciente específico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMeasurementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Medição criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Measurement"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PatientsController_findMeasurements",
        "summary": "Listar medições",
        "description": "Retorna todas as medições de um paciente específico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de medições retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Measurement"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/patients/search": {
      "get": {
        "operationId": "PatientsController_search",
        "summary": "Buscar pacientes",
        "description": "Busca pacientes por nome ou email",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Termo de busca (nome ou email)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pacientes encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/meal-plans/patient/{patientId}": {
      "get": {
        "operationId": "MealPlansController_findByPatient",
        "summary": "Buscar planos por paciente",
        "description": "Retorna todos os planos alimentares de um paciente específico",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de planos alimentares retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealPlan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/search": {
      "get": {
        "operationId": "SearchController_search",
        "summary": "Busca global por pacientes e planos alimentares",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Termo de busca (mínimo 3 caracteres)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna uma lista combinada de pacientes e planos alimentares"
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartNutri API",
    "description": "API para gestão completa de nutrição",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome completo do paciente",
            "example": "João da Silva"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do paciente"
          },
          "email": {
            "type": "string",
            "description": "Email do paciente",
            "example": "joao.silva@email.com"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do paciente",
            "example": "(11) 98765-4321"
          },
          "birthDate": {
            "type": "string",
            "description": "Data de nascimento do paciente",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "description": "Gênero do paciente",
            "enum": [
              "M",
              "F",
              "OTHER"
            ],
            "example": "M"
          },
          "status": {
            "type": "string",
            "description": "Status do paciente",
            "enum": [
              "active",
              "inactive"
            ],
            "example": "active"
          },
          "height": {
            "type": "number",
            "description": "Altura do paciente em centímetros",
            "example": 170
          },
          "weight": {
            "type": "number",
            "description": "Peso do paciente em quilogramas",
            "example": 70
          },
          "goals": {
            "description": "Objetivos do paciente",
            "example": [
              "Perda de peso",
              "Ganho de massa muscular"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "description": "Alergias do paciente",
            "example": [
              "Amendoim",
              "Lactose"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "healthConditions": {
            "description": "Condições de saúde do paciente",
            "example": [
              "Diabetes",
              "Hipertensão"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "medications": {
            "description": "Medicamentos em uso",
            "example": [
              "Metformina",
              "Losartana"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "observations": {
            "type": "string",
            "description": "Observações gerais",
            "example": "Paciente com histórico familiar de diabetes"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista responsável",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "name"
        ]
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do paciente"
          },
          "name": {
            "type": "string",
            "description": "Nome do paciente"
          },
          "email": {
            "type": "string",
            "description": "Email do paciente"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do paciente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do paciente"
          },
          "address": {
            "type": "string",
            "description": "Endereço do paciente"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "cpf",
          "phone",
          "address",
          "nutritionistId"
        ]
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do paciente"
          },
          "email": {
            "type": "string",
            "description": "Email do paciente"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do paciente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do paciente"
          },
          "address": {
            "type": "string",
            "description": "Endereço do paciente"
          }
        }
      },
      "CreateMeasurementDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Data da medição",
            "example": "2024-03-20"
          },
          "weight": {
            "type": "number",
            "description": "Peso em quilogramas",
            "example": 70.5
          },
          "bodyFat": {
            "type": "number",
            "description": "Percentual de gordura corporal",
            "example": 15.5
          },
          "muscleMass": {
            "type": "number",
            "description": "Massa muscular em quilogramas",
            "example": 35.2
          },
          "bodyWater": {
            "type": "number",
            "description": "Percentual de água corporal",
            "example": 60.5
          },
          "visceralFat": {
            "type": "number",
            "description": "Gordura visceral",
            "example": 8.5
          },
          "measurements": {
            "type": "object",
            "description": "Medidas corporais",
            "example": {
              "chest": 95,
              "waist": 80,
              "hip": 100,
              "arm": 32,
              "thigh": 55
            }
          },
          "patientId": {
            "type": "string",
            "description": "ID do paciente",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista responsável",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "date",
          "weight",
          "measurements",
          "patientId",
          "nutritionistId"
        ]
      },
      "Measurement": {
        "type": "object",
        "properties": {}
      },
...
      "Nutritionist": {
        "type": "object",
        "properties": {}
      },
      "Meal": {
        "type": "object",
        "properties": {
          "totalCalories": {
            "type": "number",
            "example": 500.25
          },
          "totalProtein": {
            "type": "number",
            "example": 30.5
          },
          "totalCarbs": {
            "type": "number",
            "example": 60.75
          },
          "totalFat": {
            "type": "number",
            "example": 20.3
          }
        },
        "required": [
          "totalCalories",
          "totalProtein",
          "totalCarbs",
          "totalFat"
        ]
      },
      "MealPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "Plano Alimentar Semanal"
          },
          "notes": {
            "type": "string",
            "example": "Plano alimentar para perda de peso"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-27"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "patientId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nutritionist": {
            "$ref": "#/components/schemas/Nutritionist"
          },
          "nutritionistId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20T10:00:00Z"
          },
          "dailyCalories": {
            "type": "number",
            "example": 2000.5
          },
          "dailyProtein": {
            "type": "number",
            "example": 150.75
          },
          "dailyCarbs": {
            "type": "number",
            "example": 250.3
          },
          "dailyFat": {
            "type": "number",
            "example": 80.2
          }
        },
        "required": [
          "id",
          "name",
          "notes",
          "startDate",
          "endDate",
          "patient",
          "patientId",
          "nutritionist",
          "nutritionistId",
          "meals",
          "createdAt",
          "updatedAt",
          "dailyCalories",
          "dailyProtein",
          "dailyCarbs",
          "dailyFat"
        ]
      },
...
    }
  }
}